refactor: split core/client.ts and MemoryGraphSection.tsx into focused modules
- Extract Zod schemas from client.ts → client-schemas.ts (159 lines)
- Extract utility/normalizer/parser functions → client-helpers.ts (359 lines)
- Extract MemoriesClientError → client-error.ts to avoid circular deps
- Extract graph types, constants, utils, canvas builder → memory-graph-helpers.ts (293 lines)
- client.ts: 1294 → 799 lines (-38%)
- MemoryGraphSection.tsx: 1649 → 1380 lines (-16%)
