+++ packages/core/src/client.ts
@@ -7,8 +7,6 @@ import type {
   BulkForgetResult,
   ContextGetInput,
   ContextGetOptions,
-  ContextMode,
-  ContextStrategy,
   ContextResult,
   ManagementKeyCreateInput,
   ManagementKeyCreateResult,
@@ -32,6 +30,43 @@ import type {
   MutationResult,
   VacuumResult,
 } from "./types"
+import {
+  apiErrorSchema,
+  baseUrlSchema,
+  bulkForgetResultSchema,
+  contextStructuredSchema,
+  managementKeyCreateSchema,
+  managementKeyRevokeSchema,
+  managementKeyStatusSchema,
+  managementTenantDisableSchema,
+  managementTenantListSchema,
+  managementTenantUpsertSchema,
+  memoriesStructuredSchema,
+  responseEnvelopeSchema,
+  skillFilesStructuredSchema,
+  vacuumResultSchema,
+} from "./client-schemas"
+import {
+  type ContextGetMethod,
+  type ParsedToolResult,
+  type RpcResponsePayload,
+  deriveMcpUrl,
+  deriveSdkBaseUrl,
+  messageFromEnvelope,
+  normalizeContextInput,
+  parseStructuredData,
+  parseToolResult,
+  pickMemoriesForMode,
+  readDefaultApiKey,
+  stripTrailingSlash,
+  toClientError,
+  toMemoryRecord,
+  toSkillFileRecord,
+  toTypedHttpError,
+} from "./client-helpers"
+import { MemoriesClientError } from "./client-error"
+
+export { MemoriesClientError } from "./client-error"
 
 export interface MemoriesClientOptions {
   apiKey?: string
@@ -41,535 +76,6 @@ export interface MemoriesClientOptions {
   tenantId?: string
   fetch?: typeof fetch
   headers?: Record<string, string>
-}
-
-interface RpcErrorPayload {
-  code: number
-  message: string
-  data?: unknown
-}
-
-interface RpcResponsePayload {
-  result?: unknown
-  error?: RpcErrorPayload
-}
-
-const baseUrlSchema = z.string().url()
-const apiErrorSchema = z.object({
-  type: z.string(),
-  code: z.string(),
-  message: z.string(),
-  status: z.number().int().optional(),
-  retryable: z.boolean().optional(),
-  details: z.unknown().optional(),
-})
-
-const responseEnvelopeSchema = z.object({
-  ok: z.boolean(),
-  data: z.unknown().nullable(),
-  error: apiErrorSchema.nullable(),
-  meta: z.record(z.string(), z.unknown()).optional(),
-})
-
-const legacyHttpErrorSchema = z.object({
-  error: z.union([z.string(), apiErrorSchema]).optional(),
-  errorDetail: apiErrorSchema.optional(),
-  message: z.string().optional(),
-})
-
-const structuredMemorySchema = z.object({
-  id: z.string().nullable().optional(),
-  content: z.string(),
-  type: z.string(),
-  layer: z.string().optional(),
-  expiresAt: z.string().nullable().optional(),
-  scope: z.string(),
-  projectId: z.string().nullable().optional(),
-  tags: z.array(z.string()).optional().default([]),
-  graph: z
-    .object({
-      whyIncluded: z.literal("graph_expansion"),
-      linkedViaNode: z.string(),
-      edgeType: z.string(),
-      hopCount: z.number().int().nonnegative(),
-      seedMemoryId: z.string(),
-    })
-    .nullable()
-    .optional(),
-})
-
-const structuredSkillFileSchema = z.object({
-  id: z.string(),
-  path: z.string(),
-  content: z.string(),
-  scope: z.string(),
-  projectId: z.string().nullable().optional(),
-  userId: z.string().nullable().optional(),
-  createdAt: z.string(),
-  updatedAt: z.string(),
-})
-
-const contextStructuredSchema = z.object({
-  rules: z.array(structuredMemorySchema).optional().default([]),
-  memories: z.array(structuredMemorySchema).optional().default([]),
-  skillFiles: z.array(structuredSkillFileSchema).optional().default([]),
-  workingMemories: z.array(structuredMemorySchema).optional().default([]),
-  longTermMemories: z.array(structuredMemorySchema).optional().default([]),
-  trace: z
-    .object({
-      requestedStrategy: z.union([z.literal("baseline"), z.literal("hybrid_graph")]).optional(),
-      strategy: z.union([z.literal("baseline"), z.literal("hybrid_graph")]),
-      graphDepth: z.union([z.literal(0), z.literal(1), z.literal(2)]),
-      graphLimit: z.number().int().nonnegative(),
-      rolloutMode: z.union([z.literal("off"), z.literal("shadow"), z.literal("canary")]).optional(),
-      shadowExecuted: z.boolean().optional(),
-      baselineCandidates: z.number().int().nonnegative(),
-      graphCandidates: z.number().int().nonnegative(),
-      graphExpandedCount: z.number().int().nonnegative(),
-      fallbackTriggered: z.boolean().optional(),
-      fallbackReason: z.string().nullable().optional(),
-      totalCandidates: z.number().int().nonnegative(),
-    })
-    .optional(),
-})
-
-const memoriesStructuredSchema = z.object({
-  memories: z.array(structuredMemorySchema).optional().default([]),
-})
-
-const skillFilesStructuredSchema = z.object({
-  skillFiles: z.array(structuredSkillFileSchema).optional().default([]),
-  count: z.number().int().nonnegative().optional(),
-})
-
-const mutationEnvelopeDataSchema = z.object({
-  message: z.string().optional(),
-})
-
-const bulkForgetResultSchema = z.object({
-  count: z.number().int().nonnegative(),
-  ids: z.array(z.string()).optional(),
-  memories: z.array(z.object({
-    id: z.string(),
-    type: z.string(),
-    contentPreview: z.string(),
-  })).optional(),
-  message: z.string(),
-})
-
-const vacuumResultSchema = z.object({
-  purged: z.number().int().nonnegative(),
-  message: z.string(),
-})
-
-const managementKeyStatusSchema = z.object({
-  hasKey: z.boolean(),
-  keyPreview: z.string().optional(),
-  createdAt: z.string().nullable().optional(),
-  expiresAt: z.string().nullable().optional(),
-  isExpired: z.boolean().optional(),
-})
-
-const managementKeyCreateSchema = z.object({
-  apiKey: z.string(),
-  keyPreview: z.string().optional(),
-  createdAt: z.string().optional(),
-  expiresAt: z.string().optional(),
-  message: z.string().optional(),
-})
-
-const managementKeyRevokeSchema = z.object({
-  ok: z.boolean(),
-})
-
-const managementTenantSchema = z.object({
-  tenantId: z.string(),
-  tursoDbUrl: z.string(),
-  tursoDbName: z.string().nullable().optional(),
-  status: z.string(),
-  metadata: z.record(z.string(), z.unknown()).optional(),
-  createdAt: z.string().optional(),
-  updatedAt: z.string().optional(),
-  lastVerifiedAt: z.string().nullable().optional(),
-})
-
-const managementTenantListSchema = z.object({
-  tenantDatabases: z.array(managementTenantSchema),
-  count: z.number().int(),
-})
-
-const managementTenantUpsertSchema = z.object({
-  tenantDatabase: managementTenantSchema,
-  provisioned: z.boolean(),
-  mode: z.string(),
-})
-
-const managementTenantDisableSchema = z.object({
-  ok: z.boolean(),
-  tenantId: z.string(),
-  status: z.string(),
-  updatedAt: z.string(),
-})
-
-function errorTypeForStatus(status: number): MemoriesErrorData["type"] {
-  if (status === 400) return "validation_error"
-  if (status === 401 || status === 403) return "auth_error"
-  if (status === 404) return "not_found_error"
-  if (status === 429) return "rate_limit_error"
-  if (status >= 500) return "internal_error"
-  return "http_error"
-}
-
-function isRetryableStatus(status: number): boolean {
-  return status === 429 || status >= 500
-}
-
-function toTypedHttpError(status: number, payload: unknown): MemoriesErrorData {
-  const parsed = legacyHttpErrorSchema.safeParse(payload)
-  if (!parsed.success) {
-    return {
-      type: errorTypeForStatus(status),
-      code: `HTTP_${status}`,
-      message: `HTTP ${status}`,
-      status,
-      retryable: isRetryableStatus(status),
-      details: payload,
-    }
-  }
-
-  if (parsed.data.errorDetail) {
-    return {
-      ...parsed.data.errorDetail,
-      status: parsed.data.errorDetail.status ?? status,
-      retryable: parsed.data.errorDetail.retryable ?? isRetryableStatus(status),
-    }
-  }
-
-  if (typeof parsed.data.error !== "string" && parsed.data.error) {
-    return {
-      ...parsed.data.error,
-      status: parsed.data.error.status ?? status,
-      retryable: parsed.data.error.retryable ?? isRetryableStatus(status),
-    }
-  }
-
-  const message = typeof parsed.data.error === "string"
-    ? parsed.data.error
-    : parsed.data.message ?? `HTTP ${status}`
-
-  return {
-    type: errorTypeForStatus(status),
-    code: `HTTP_${status}`,
-    message,
-    status,
-    retryable: isRetryableStatus(status),
-  }
-}
-
-export class MemoriesClientError extends Error {
-  readonly status?: number
-  readonly code?: number
-  readonly data?: unknown
-  readonly type?: MemoriesErrorData["type"]
-  readonly errorCode?: string
-  readonly retryable?: boolean
-  readonly details?: unknown
-
-  constructor(
-    message: string,
-    options?: {
-      status?: number
-      code?: number
-      data?: unknown
-      type?: MemoriesErrorData["type"]
-      errorCode?: string
-      retryable?: boolean
-      details?: unknown
-    }
-  ) {
-    super(message)
-    this.name = "MemoriesClientError"
-    this.status = options?.status
-    this.code = options?.code
-    this.type = options?.type
-    this.errorCode = options?.errorCode
-    this.retryable = options?.retryable
-    this.details = options?.details
-    this.data = options?.data ?? options?.details
-  }
-}
-
-function readDefaultApiKey(): string | undefined {
-  if (typeof process === "undefined") return undefined
-  return process.env.MEMORIES_API_KEY
-}
-
-function normalizeMemoryType(type: string): MemoryRecord["type"] {
-  if (type === "rule" || type === "decision" || type === "fact" || type === "note" || type === "skill") {
-    return type
-  }
-  return "note"
-}
-
-function normalizeMemoryLayer(layer: string | undefined, type: MemoryRecord["type"]): MemoryRecord["layer"] {
-  if (layer === "rule" || layer === "working" || layer === "long_term") {
-    return layer
-  }
-  return type === "rule" ? "rule" : "long_term"
-}
-
-function normalizeMemoryScope(scope: string): MemoryRecord["scope"] {
-  if (scope === "global" || scope === "project" || scope === "unknown") {
-    return scope
-  }
-  return "unknown"
-}
-
-function toMemoryRecord(memory: z.infer<typeof structuredMemorySchema>): MemoryRecord {
-  const type = normalizeMemoryType(memory.type)
-  return {
-    id: memory.id ?? null,
-    content: memory.content,
-    type,
-    layer: normalizeMemoryLayer(memory.layer, type),
-    expiresAt: memory.expiresAt ?? null,
-    scope: normalizeMemoryScope(memory.scope),
-    projectId: memory.projectId ?? null,
-    tags: memory.tags ?? [],
-    graph: memory.graph ?? null,
-  }
-}
-
-function toSkillFileRecord(skillFile: z.infer<typeof structuredSkillFileSchema>): SkillFileRecord {
-  return {
-    id: skillFile.id,
-    path: skillFile.path,
-    content: skillFile.content,
-    scope: normalizeMemoryScope(skillFile.scope),
-    projectId: skillFile.projectId ?? null,
-    userId: skillFile.userId ?? null,
-    createdAt: skillFile.createdAt,
-    updatedAt: skillFile.updatedAt,
-  }
-}
-
-function dedupeMemories(records: z.infer<typeof structuredMemorySchema>[]): z.infer<typeof structuredMemorySchema>[] {
-  const seen = new Set<string>()
-  const deduped: z.infer<typeof structuredMemorySchema>[] = []
-
-  for (const memory of records) {
-    const key = memory.id ?? `${memory.type}:${memory.scope}:${memory.content}`
-    if (seen.has(key)) continue
-    seen.add(key)
-    deduped.push(memory)
-  }
-
-  return deduped
-}
-
-const contextModes = new Set<ContextMode>(["all", "working", "long_term", "rules_only"])
-const contextStrategies = new Set<ContextStrategy>(["baseline", "hybrid_graph"])
-
-interface ContextBuckets {
-  working: z.infer<typeof structuredMemorySchema>[]
-  longTerm: z.infer<typeof structuredMemorySchema>[]
-}
-
-interface NormalizedContextInput extends ContextGetInput {
-  mode: ContextMode
-  strategy: ContextStrategy
-  graphDepth: 0 | 1 | 2
-  graphLimit: number
-}
-
-type ContextGetMethod = {
-  (input?: ContextGetInput): Promise<ContextResult>
-  (query?: string, options?: ContextGetOptions): Promise<ContextResult>
-}
-
-function normalizeContextMode(mode: unknown): ContextMode {
-  if (typeof mode === "string" && contextModes.has(mode as ContextMode)) {
-    return mode as ContextMode
-  }
-  return "all"
-}
-
-function normalizeContextStrategy(strategy: unknown): ContextStrategy {
-  if (typeof strategy === "string" && contextStrategies.has(strategy as ContextStrategy)) {
-    return strategy as ContextStrategy
-  }
-  return "baseline"
-}
-
-function normalizeGraphDepth(depth: unknown): 0 | 1 | 2 {
-  if (depth === 0 || depth === 1 || depth === 2) {
-    return depth
-  }
-  return 1
-}
-
-function normalizeGraphLimit(limit: unknown): number {
-  if (typeof limit !== "number" || !Number.isFinite(limit)) {
-    return 8
-  }
-  return Math.max(1, Math.min(Math.floor(limit), 50))
-}
-
-function normalizeContextInput(
-  inputOrQuery?: string | ContextGetInput,
-  options: ContextGetOptions = {}
-): NormalizedContextInput {
-  const fromObject =
-    inputOrQuery && typeof inputOrQuery === "object"
-      ? inputOrQuery
-      : { query: typeof inputOrQuery === "string" ? inputOrQuery : undefined, ...options }
-
-  return {
-    ...fromObject,
-    mode: normalizeContextMode(fromObject.mode),
-    strategy: normalizeContextStrategy(fromObject.strategy),
-    graphDepth: normalizeGraphDepth(fromObject.graphDepth),
-    graphLimit: normalizeGraphLimit(fromObject.graphLimit),
-  }
-}
-
-function partitionByLayer(records: z.infer<typeof structuredMemorySchema>[]): ContextBuckets {
-  const working: z.infer<typeof structuredMemorySchema>[] = []
-  const longTerm: z.infer<typeof structuredMemorySchema>[] = []
-
-  for (const record of records) {
-    const type = normalizeMemoryType(record.type)
-    const layer = normalizeMemoryLayer(record.layer, type)
-    if (layer === "working") {
-      working.push(record)
-      continue
-    }
-    if (layer === "long_term") {
-      longTerm.push(record)
-    }
-  }
-
-  return { working, longTerm }
-}
-
-function pickMemoriesForMode(
-  mode: ContextMode,
-  structured: z.infer<typeof contextStructuredSchema>
-): z.infer<typeof structuredMemorySchema>[] {
-  const fallbackBuckets = partitionByLayer(structured.memories)
-  const working = dedupeMemories([...structured.workingMemories, ...fallbackBuckets.working])
-  const longTerm = dedupeMemories([...structured.longTermMemories, ...fallbackBuckets.longTerm])
-
-  if (mode === "rules_only") {
-    return []
-  }
-  if (mode === "working") {
-    return working
-  }
-  if (mode === "long_term") {
-    return longTerm
-  }
-  return dedupeMemories([...working, ...longTerm])
-}
-
-interface ParsedToolResult {
-  raw: string
-  structured: unknown
-  envelope: MemoriesResponseEnvelope<unknown> | null
-}
-
-function toClientError(error: MemoriesErrorData, options?: { status?: number; rpcCode?: number }) {
-  return new MemoriesClientError(error.message, {
-    status: error.status ?? options?.status,
-    code: options?.rpcCode,
-    type: error.type,
-    errorCode: error.code,
-    retryable: error.retryable,
-    details: error.details,
-  })
-}
-
-function parseToolResult(result: unknown): ParsedToolResult {
-  const parsed = z
-    .object({
-      content: z.array(z.object({ type: z.string(), text: z.string().optional() })).optional(),
-      structuredContent: z.unknown().optional(),
-    })
-    .passthrough()
-    .safeParse(result)
-
-  if (!parsed.success) {
-    return { raw: "", structured: null, envelope: null }
-  }
-
-  const textChunk = parsed.data.content?.find((entry) => entry.type === "text" && entry.text)
-  const envelope = responseEnvelopeSchema.safeParse(parsed.data.structuredContent)
-
-  if (envelope.success) {
-    const parsedEnvelope: MemoriesResponseEnvelope<unknown> = {
-      ok: envelope.data.ok,
-      data: envelope.data.data,
-      error: envelope.data.error,
-      meta: envelope.data.meta,
-    }
-    return {
-      raw: textChunk?.text ?? "",
-      structured: parsedEnvelope.data,
-      envelope: parsedEnvelope,
-    }
-  }
-
-  return {
-    raw: textChunk?.text ?? "",
-    structured: parsed.data.structuredContent ?? null,
-    envelope: null,
-  }
-}
-
-function messageFromEnvelope(envelope: MemoriesResponseEnvelope<unknown> | null): string | null {
-  if (!envelope?.data || typeof envelope.data !== "object") return null
-  const parsed = mutationEnvelopeDataSchema.safeParse(envelope.data)
-  if (!parsed.success || typeof parsed.data.message !== "string") return null
-  return parsed.data.message
-}
-
-function parseStructuredData<T>(
-  schema: z.ZodType<T>,
-  endpoint: string,
-  structured: unknown
-): T {
-  const parsed = schema.safeParse(structured)
-  if (parsed.success) {
-    return parsed.data
-  }
-
-  throw new MemoriesClientError(`Invalid SDK response data for ${endpoint}`, {
-    type: "http_error",
-    errorCode: "INVALID_RESPONSE_DATA",
-    retryable: false,
-    details: structured,
-  })
-}
-
-function stripTrailingSlash(value: string): string {
-  return value.endsWith("/") ? value.slice(0, -1) : value
-}
-
-function deriveSdkBaseUrl(baseUrl: string): string {
-  if (baseUrl.endsWith("/api/mcp")) {
-    return baseUrl.slice(0, -"/api/mcp".length)
-  }
-  if (baseUrl.endsWith("/api/sdk/v1")) {
-    return baseUrl.slice(0, -"/api/sdk/v1".length)
-  }
-  return baseUrl
-}
-
-function deriveMcpUrl(baseUrl: string, sdkBaseUrl: string): string {
-  if (baseUrl.endsWith("/api/mcp")) {
-    return baseUrl
-  }
-  return `${sdkBaseUrl}/api/mcp`
 }
 
 export class MemoriesClient {
@@ -821,7 +327,6 @@ export class MemoriesClient {
     },
 
     bulkForget: async (filters: BulkForgetFilter, options?: { dryRun?: boolean }): Promise<BulkForgetResult> => {
-      // Scope carries tenantId/userId for routing; projectId is a query-level filter only
       const rawScope = this.withDefaultScopeSdk({})
       const sdkScope = rawScope && Object.keys(rawScope).length > 0 ? rawScope : undefined
       const dryRun = options?.dryRun ?? false
